+++ addOns/pscanrulesBeta/src/main/java/org/zaproxy/zap/extension/pscanrulesBeta/InformationDisclosureSuspiciousComments.java
@@ -45,17 +45,21 @@ public class InformationDisclosureSuspiciousComments extends PluginPassiveScanne
     private static final int PLUGIN_ID = 10027;
 
     private PassiveScanThread parent = null;
-    private static final String databaseErrorFile = "xml/suspicious-comments.txt";
+    public static final String suspiciousCommentsListDir = "xml";
+    public static final String suspiciousCommentsListFile = "suspicious-comments.txt";
     private static final Logger logger =
             Logger.getLogger(InformationDisclosureSuspiciousComments.class);
 
-    private List<Pattern> patterns = null;
+    private static List<Pattern> patterns = null;
 
     @Override
     public void scanHttpRequestSend(HttpMessage msg, int id) {}
 
     @Override
     public void scanHttpResponseReceive(HttpMessage msg, int id, Source source) {
+
+        List<Pattern> patterns = getPatterns();
+
         if (msg.getResponseBody().length() > 0 && msg.getResponseHeader().isText()) {
             StringBuilder todoComments = new StringBuilder();
 
@@ -63,7 +67,7 @@ public class InformationDisclosureSuspiciousComments extends PluginPassiveScanne
                 // Just treat as text
                 String[] lines = msg.getResponseBody().toString().split("\n");
                 for (String line : lines) {
-                    for (Pattern pattern : this.getPatterns()) {
+                    for (Pattern pattern : patterns) {
                         if (pattern.matcher(line).find()) {
                             todoComments.append(line);
                             todoComments.append("\n");
@@ -78,7 +82,7 @@ public class InformationDisclosureSuspiciousComments extends PluginPassiveScanne
                 List<Tag> tags = source.getAllTags(StartTagType.COMMENT);
                 for (Tag tag : tags) {
                     String tagStr = tag.toString();
-                    for (Pattern pattern : this.getPatterns()) {
+                    for (Pattern pattern : patterns) {
                         if (pattern.matcher(tagStr).find()) {
                             todoComments.append(tagStr);
                             todoComments.append("\n");
@@ -91,7 +95,7 @@ public class InformationDisclosureSuspiciousComments extends PluginPassiveScanne
                 int offset = 0;
                 while ((el = source.getNextElement(offset, HTMLElementName.SCRIPT)) != null) {
                     String elStr = el.toString();
-                    for (Pattern pattern : this.getPatterns()) {
+                    for (Pattern pattern : patterns) {
                         if (pattern.matcher(elStr).find()) {
                             todoComments.append(elStr);
                             todoComments.append("\n");
@@ -125,34 +129,41 @@ public class InformationDisclosureSuspiciousComments extends PluginPassiveScanne
         parent.raiseAlert(id, alert);
     }
 
-    private List<Pattern> getPatterns() {
+    private static List<Pattern> getPatterns() {
         if (patterns == null) {
             patterns = new ArrayList<>();
-            String line = null;
-            File f = new File(Constant.getZapHome() + File.separator + databaseErrorFile);
-            BufferedReader reader = null;
+
             try {
-                reader = new BufferedReader(new FileReader(f));
-                while ((line = reader.readLine()) != null) {
-                    if (!line.startsWith("#") && line.length() > 0) {
-                        patterns.add(
-                                Pattern.compile("\\b" + line + "\\b", Pattern.CASE_INSENSITIVE));
+                File f =
+                        new File(
+                                Constant.getZapHome()
+                                        + File.separator
+                                        + suspiciousCommentsListDir
+                                        + File.separator
+                                        + suspiciousCommentsListFile);
+                if (!f.exists()) {
+                    throw new IOException("Couldn't find resource: " + f.getAbsolutePath());
+                }
+                try (BufferedReader reader = new BufferedReader(new FileReader(f))) {
+                    String line = null;
+                    while ((line = reader.readLine()) != null) {
+                        line = line.trim();
+                        if (!line.startsWith("#") && line.length() > 0) {
+                            patterns.add(
+                                    Pattern.compile(
+                                            "\\b" + line + "\\b", Pattern.CASE_INSENSITIVE));
+                        }
                     }
                 }
             } catch (IOException e) {
                 logger.error(
-                        "Error on opening/reading database error file. File: "
-                                + f.getAbsolutePath()
+                        "Error on opening/reading suspicious comments file: "
+                                + File.separator
+                                + suspiciousCommentsListDir
+                                + File.separator
+                                + suspiciousCommentsListFile
                                 + " Error: "
                                 + e.getMessage());
-            } finally {
-                if (reader != null) {
-                    try {
-                        reader.close();
-                    } catch (IOException e) {
-                        logger.debug("Error on closing the file reader. Error: " + e.getMessage());
-                    }
-                }
             }
         }
         return patterns;