+++ addOns/pscanrulesBeta/CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+- Tweak Information Disclosure - Suspicious Comments scanner to ignore whitespace before/after suspicious comments terms in the suspicious-comments.txt config file.
 - Only scan for Servlet Parameter Pollution at LOW threshold (part of Issue 4454).
 
 ## [19] - 2019-06-07